LPS-128873 Migrate customized property accessors: PropertyAccessor -> PropertyAccessStrategy that works as factory of PropertyAccess
