+++ portal-impl/src/com/liferay/portal/dao/orm/hibernate/CamelCasePropertyAccessor.java
@@ -14,29 +14,44 @@
 
 package com.liferay.portal.dao.orm.hibernate;
 
+import org.hibernate.property.access.spi.PropertyAccess;
+
 /**
- * @author Brian Wing Shun Chan
+ * @author Dante Wang
  */
 public class CamelCasePropertyAccessor extends LiferayPropertyAccessor {
 
+	public static final CamelCasePropertyAccessor INSTANCE =
+		new CamelCasePropertyAccessor();
+
 	@Override
-	protected String formatPropertyName(String propertyName) {
-		if (propertyName.length() < 3) {
-			return super.formatPropertyName(propertyName);
-		}
+	public PropertyAccess buildPropertyAccess(
+		Class containerJavaType, String propertyName) {
 
-		char c0 = propertyName.charAt(0);
-		char c1 = propertyName.charAt(1);
-		char c2 = propertyName.charAt(2);
+		return new LiferayPropertyAccess(
+			this, containerJavaType, propertyName) {
 
-		if (Character.isLowerCase(c0) && Character.isUpperCase(c1) &&
-			Character.isLowerCase(c2)) {
+			@Override
+			protected String formatPropertyName(String propertyName) {
+				if (propertyName.length() < 3) {
+					return super.formatPropertyName(propertyName);
+				}
 
-			propertyName =
-				Character.toUpperCase(c0) + propertyName.substring(1);
-		}
+				char c0 = propertyName.charAt(0);
+				char c1 = propertyName.charAt(1);
+				char c2 = propertyName.charAt(2);
 
-		return super.formatPropertyName(propertyName);
+				if (Character.isLowerCase(c0) && Character.isUpperCase(c1) &&
+					Character.isLowerCase(c2)) {
+
+					propertyName =
+						Character.toUpperCase(c0) + propertyName.substring(1);
+				}
+
+				return super.formatPropertyName(propertyName);
+			}
+
+		};
 	}
 
 }
