+++ portal-impl/src/com/liferay/portal/dao/orm/hibernate/PrivatePropertyAccessor.java
@@ -17,45 +17,35 @@ package com.liferay.portal.dao.orm.hibernate;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.impl.BaseModelImpl;
 
-import org.hibernate.PropertyNotFoundException;
-import org.hibernate.property.DirectPropertyAccessor;
-import org.hibernate.property.Getter;
-import org.hibernate.property.Setter;
+import org.hibernate.property.access.internal.PropertyAccessFieldImpl;
+import org.hibernate.property.access.internal.PropertyAccessStrategyFieldImpl;
+import org.hibernate.property.access.spi.PropertyAccess;
 
 /**
- * @author Shuyang Zhou
+ * @author Dante Wang
  */
 @SuppressWarnings("rawtypes")
-public class PrivatePropertyAccessor extends DirectPropertyAccessor {
+public class PrivatePropertyAccessor extends PropertyAccessStrategyFieldImpl {
+
+	public static final PrivatePropertyAccessor INSTANCE =
+		new PrivatePropertyAccessor();
 
 	@Override
-	public Getter getGetter(Class clazz, String propertyName)
-		throws PropertyNotFoundException {
+	public PropertyAccess buildPropertyAccess(
+		Class containerJavaType, String propertyName) {
 
 		Class<?> superClass = null;
 
-		while ((superClass = clazz.getSuperclass()) != BaseModelImpl.class) {
-			clazz = superClass;
+		while ((superClass = containerJavaType.getSuperclass()) !=
+					BaseModelImpl.class) {
+
+			containerJavaType = superClass;
 		}
 
 		propertyName = StringPool.UNDERLINE.concat(propertyName);
 
-		return super.getGetter(clazz, propertyName);
-	}
-
-	@Override
-	public Setter getSetter(Class clazz, String propertyName)
-		throws PropertyNotFoundException {
-
-		Class<?> superClass = null;
-
-		while ((superClass = clazz.getSuperclass()) != BaseModelImpl.class) {
-			clazz = superClass;
-		}
-
-		propertyName = StringPool.UNDERLINE.concat(propertyName);
-
-		return super.getSetter(clazz, propertyName);
+		return new PropertyAccessFieldImpl(
+			this, containerJavaType, propertyName);
 	}
 
 }
